commOccu objects

commOccu objects

Value

commOccu object

Details

The data slot is a list of model input data. While the exact content depends on function input, it can be summarized as:

y	array of detection histories. Dimensions are: y[species, station, occasion]

effort_binary	matrix of binary (1/0) survey effort. Only used to ensure p = 0 when effort = 0. Dimensions are: effort_binary[station, occasion]

site-occasion covariates	The content of data_list$obsCovs as named matrices with dimensions [station, occasion]

site covariates	The columns of data_list$siteCovs as named vectors (length = number of stations)

M	Number of species

J	Number of stations

maxocc	Number of occasions

Slots

modelText

JAGS model code as a character vector (made up of code chunks, use cat() to print)

params

Parameters to monitor in the model runs

inits_fun

Function to create start values for the MCMC chains. It being a function ensures different values in each chain

data

List with data needed to run the model (detection & effort matrices, site covariates, number of species / stations / occasions)

input

Input data_list (unchanged)

nimble

logical indicator for whether it is a Nimble model

modelFile

Path of the text file containing the model code