Summarize community occupancy model

Gives an overview of the number of species, stations and occasions in a commOccu object. Also returns covariates.

# S4 method for commOccu
summary(object, ...)

Arguments

object
commOccu object

...
currently ignored

Value

Model summary printed to console

Details

The summary method is very basic and still work in progress.