Create a report about a camera trapping survey and species detections

This function creates a report about a camera trapping survey and species records. It uses a camera trap station information table and a record table (generated with recordTable) as input. Output tables can be saved and a zip file for simple data sharing can be created easily.

surveyReport (recordTable,
  CTtable,
  speciesCol = "Species",
  stationCol = "Station",
  cameraCol,
  setupCol,
  retrievalCol,
  CTDateFormat = "%Y-%m-%d",
  CTHasProblems = FALSE,
  recordDateTimeCol = "DateTimeOriginal",
  recordDateTimeFormat = "%Y-%m-%d %H:%M:%S",
  Xcol,
  Ycol,
  sinkpath,
  makezip
)

Arguments

recordTable
data.frame containing a species record table as given by recordTable

CTtable
data.frame containing information about location and trapping period of camera trap stations (equivalent to camtraps

speciesCol
character. name of the column specifying Species ID in recordTable

stationCol
character. name of the column specifying Station ID in CTtable and recordTable

cameraCol
character. name of the column specifying Camera ID in CTtable and recordTable

setupCol
character. name of the column containing camera setup dates in CTtable

retrievalCol
character. name of the column containing camera retrieval dates in CTtable

CTDateFormat
character. The format of columns setupCol and retrievalCol (and potential problem columns) in CTtable. Must be interpretable by either as.Date or the "orders" argument parse_date_time in lubridate.

CTHasProblems
logical. Are there periods of camera malfunction specified in CTtable?

recordDateTimeCol
character. The name of the column containing date and time of records in recordTable

recordDateTimeFormat
character. The date/time format of column recordDateTimeCol in recordTable.

Xcol
character. name of the column specifying x coordinates in CTtable. Used to create detection maps if makezip is TRUE. (optional)

Ycol
character. name of the column specifying y coordinates in CTtable. Used to create detection maps if makezip is TRUE. (optional)

sinkpath
character. The directory into which the survey report is saved (optional)

makezip
logical. Create a zip file containing tables, plots and maps in sinkpath?

Details

dateFormat defaults to "YYYY-MM-DD", e.g. "2014-10-31". It can be specified either in the format required by strptime or the 'orders' argument in parse_date_time in lubridate. In the example above, "YYYY-MM-DD" would be specified as "%Y-%m-%d" or "ymd". If CTHasProblems is set to TRUE, the function tries to find columns ProblemX_from and ProblemX_to in CTtable (X designates numbers from 1 to n in which a camera or station was not operational). If there are no such columns all stations are assumed to have been operational uninterruptedly from setup to retrieval.

recordDateTimeFormat defaults to the "YYYY-MM-DD HH:MM:SS" convention, e.g. "2014-09-30 22:59:59". recordDateTimeFormat can be interpreted either by base-R via strptime or in lubridate via parse_date_time (argument "orders"). lubridate will be used if there are no "%" characters in recordDateTimeFormat.

For "YYYY-MM-DD HH:MM:SS", recordDateTimeFormat would be either "%Y-%m-%d %H:%M:%S" or "ymd HMS". For details on how to specify date and time formats in R see strptime or parse_date_time.

Value

An invisible list containing 5 data.frames.

survey_dates

station and image date ranges, number of total and active trap nights, number of cameras per station

species_by_station

species numbers by station

events_by_species

number of events and stations by species

events_by_station

number of events for every species by station (only species that were recorded)

events_by_station2

number of events for all species at all stations (including species that were not recorded)

The output will be saved to a .txt file if sinkpath is defined. If makezip is TRUE, a zip file will be created in sinkpath. It contains single-species activity plots, detection maps (if Xcol and Ycol are defined), the survey report tables, the record table and the camera trap station table, and an example R script.
See also

Examples

data(camtraps)
data(recordTableSample)

reportTest <- surveyReport (recordTable          = recordTableSample,
                            CTtable              = camtraps,
                            speciesCol           = "Species",
                            stationCol           = "Station",
                            setupCol             = "Setup_date",
                            retrievalCol         = "Retrieval_date",
                            CTDateFormat         = "%d/%m/%Y",
                            recordDateTimeCol    = "DateTimeOriginal",
                            recordDateTimeFormat = "%Y-%m-%d %H:%M:%S")

#> 
#> -------------------------------------------------------
#> [1] "Total number of stations:  3"
#> 
#> -------------------------------------------------------
#> [1] "Number of operational stations:  3"
#> 
#> -------------------------------------------------------
#> [1] "n nights with cameras set up (operational or not. NOTE: only correct if 1 camera per station): 128"
#> 
#> -------------------------------------------------------
#> [1] "n nights with cameras set up and active (trap nights. NOTE: only correct if 1 camera per station): 128"
#> 
#> -------------------------------------------------------
#> [1] "total trapping period:  2009-04-02 - 2009-05-17"

class(reportTest)  # a list with

#> [1] "list"

length(reportTest) # 5 elements

#> [1] 5

reportTest[[1]]    # camera trap operation times and image date ranges

#>    Station setup_date first_image_date last_image_date retrieval_date
#> 1 StationA 2009-04-02       2009-04-10      2009-05-07     2009-05-14
#> 2 StationB 2009-04-03       2009-04-05      2009-05-14     2009-05-16
#> 3 StationC 2009-04-04       2009-04-06      2009-05-12     2009-05-17
#>   n_nights_total n_nights_active n_cameras
#> 1             42              42         1
#> 2             43              43         1
#> 3             43              43         1

reportTest[[2]]    # number of species by station

#>    Station n_species
#> 1 StationA         2
#> 2 StationB         3
#> 3 StationC         4

reportTest[[3]]    # number of events and number of stations by species

#>   species n_events n_stations
#> 1     EGY        6          1
#> 2     MNE        2          1
#> 3     PBE       18          3
#> 4     TRA        8          1
#> 5     VTA        5          3

reportTest[[4]]    # number of species events by station

#>    Station Species n_events
#> 1 StationA     PBE        4
#> 2 StationA     VTA        2
#> 3 StationB     MNE        2
#> 4 StationB     PBE        8
#> 5 StationB     VTA        2
#> 6 StationC     EGY        6
#> 7 StationC     PBE        6
#> 8 StationC     TRA        8
#> 9 StationC     VTA        1

reportTest[[5]]    # number of species events by station including 0s (non-observed species)

#>     Station Species n_events
#> 1  StationA     EGY        0
#> 2  StationA     MNE        0
#> 3  StationA     PBE        4
#> 4  StationA     TRA        0
#> 5  StationA     VTA        2
#> 6  StationB     EGY        0
#> 7  StationB     MNE        2
#> 8  StationB     PBE        8
#> 9  StationB     TRA        0
#> 10 StationB     VTA        2
#> 11 StationC     EGY        6
#> 12 StationC     MNE        0
#> 13 StationC     PBE        6
#> 14 StationC     TRA        8
#> 15 StationC     VTA        1

# with camera problems

reportTest_problem <- surveyReport (recordTable          = recordTableSample,
                                    CTtable              = camtraps,
                                    speciesCol           = "Species",
                                    stationCol           = "Station",
                                    setupCol             = "Setup_date",
                                    retrievalCol         = "Retrieval_date",
                                    CTDateFormat         = "%d/%m/%Y",
                                    recordDateTimeCol    = "DateTimeOriginal",
                                    recordDateTimeFormat = "%Y-%m-%d %H:%M:%S",
                                    CTHasProblems        = TRUE)

#> 
#> -------------------------------------------------------
#> [1] "Total number of stations:  3"
#> 
#> -------------------------------------------------------
#> [1] "Number of operational stations:  3"
#> 
#> -------------------------------------------------------
#> [1] "n nights with cameras set up (operational or not. NOTE: only correct if 1 camera per station): 128"
#> 
#> -------------------------------------------------------
#> [1] "n nights with cameras set up and active (trap nights. NOTE: only correct if 1 camera per station): 123"
#> 
#> -------------------------------------------------------
#> [1] "total trapping period:  2009-04-02 - 2009-05-17"

reportTest_problem$survey_dates

#>    Station setup_date first_image_date last_image_date retrieval_date
#> 1 StationA 2009-04-02       2009-04-10      2009-05-07     2009-05-14
#> 2 StationB 2009-04-03       2009-04-05      2009-05-14     2009-05-16
#> 3 StationC 2009-04-04       2009-04-06      2009-05-12     2009-05-17
#>   n_nights_total n_nights_active n_cameras
#> 1             42              42         1
#> 2             43              43         1
#> 3             43              38         1

if (FALSE) {
# run again with sinkpath defined
reportTest <- surveyReport (recordTable          = recordTableSample,
                            CTtable              = camtraps,
                            speciesCol           = "Species",
                            stationCol           = "Station",
                            setupCol             = "Setup_date",
                            retrievalCol         = "Retrieval_date",
                            CTDateFormat         = "%d/%m/%Y",
                            recordDateTimeCol    = "DateTimeOriginal",
                            recordDateTimeFormat = "%Y-%m-%d %H:%M:%S",
                            sinkpath             = getwd())

# have a look at the text file
readLines(list.files(getwd(), pattern = paste("survey_report_", Sys.Date(), ".txt", sep = ""),
                                         full.names = TRUE))
}