Overview of the functions in the camtrapR package

This package provides a streamlined workflow for processing data generated in camera trap-based wildlife studies and prepares input for further analyses, particularly in occupancy and spatial capture-recapture frameworks. It suggests a simple data structure and provides functions for managing digital camera trap photographs (and videos), generating record tables, maps of species richness and species detections and species activity diagrams. It further helps prepare subsequent analyses by creating detection/non-detection matrices for occupancy analyses, e.g. in the unmarked or ubms packages, and capthist objects for spatial capture-recapture analyses in the secr package. In addition, basic survey statistics are computed. The functions build on one another in a logical sequence. The only manual input needed is species (and individual) identification, which is achieved by moving images into species directories or by tagging images in image management software. Besides, a table holding basic information about camera trap station IDs, locations and trapping periods must be created in spreadsheet software.

User support

The camtrapR Google group is an online support and help forum for camtrapR users. You can find it here: https://groups.google.com/forum/#!forum/camtrapr.

Details

Image metadata (such as date and time or user-assigned tags) are extracted from the images using Phil Harvey's ExifTool (available from https://exiftool.org/) and the information is stored in a record table. An adjustable criterion for temporal independence of records can be applied. Maps of species presence and species richness can be generated. Several functions are available for plotting single- and two-species activity patterns. Information about the camera-specific trapping periods (and periods of malfunction) are summarized into a matrix about camera trap operability. These, together with the record table, are used to generate species detection histories for occupancy and spatial capture-recapture analyses. The user has considerable freedom in generating the detection histories; sampling occasion length, beginning date and and occasion start times are adjustable. In addition, trapping effort (i.e. active trap nights per station and occasion) can be computed for use as a covariate / offset on detection probability.

Image organisation and management

The functions in this section set up a directory structure for storing camera trap images and identifying species and individuals from images. They build on one another and can be run in sequential order as needed.

createStationFolders	Create camera trap station directories for raw images

fixDateTimeOriginal	Fix DateTimeOriginal Exif metadata tag in Reconyx Hyperfire cameras

OCRdataFields	Optical character recognition (OCR) from data fields in images

writeDateTimeOriginal	Write values to DateTimeOriginal tag in image metadata

timeShiftImages	Apply time shifts to JPEG images

imageRename	Copy and rename images based on station ID and image creation date

addCopyrightTag	Write a copyright tag into JPEG image metadata

appendSpeciesNames	Add or remove species names from image filenames

Species / individual identification

These functions assist in species identification and prepare individual identification of animals.

checkSpeciesNames	Check species names against the ITIS taxonomic database

createSpeciesFolders	Create directories for species identification

checkSpeciesIdentification	Consistency check on species image identification

getSpeciesImages	Gather all images of a species in a new directory

Image data extraction

These function use the directory structure built above (Section 'Image management workflow') and a table containing basic information about camera traps and/or stations (IDs, location, trapping period).

recordTable	Create a species record table from camera trap images and videos

recordTableIndividual	Create a single-species record table from camera trap images and videos with individual IDs

exifTagNames	Return Exif metadata tags and tag names from JPEG images

addToPath	Add the directory containing exiftool.exe to PATH temporarily (Windows only)

filterRecordTable	Filter existing record table for temporal independence.

Data exploration and visualisation

These plots are generated from the record table and the camera trap table.

detectionMaps	Generate maps of species richness and species presence by station, export shapefiles

activityHistogram	Single-species diel activity histograms

activityDensity	Single-species diel activity kernel density estimation plots

activityRadial	Single-species diel activity radial plot

activityOverlap	Two-species diel activity overlap plots and estimates

Community (multi-species) occupancy models

communityModel	Create a community (multi-species) occupancy model for JAGS or Nimble

commOccu-class	commOccu objects

fit,commOccu-method	Fit a community (multi-species) occupancy model

predict,commOccu-method	Predictions from community occupancy models

summary,commOccu-method	Summarize community occupancy model

plot_coef	Plot effect sizes of covariates in community occupancy model

plot_effects	Plot Marginal Effects of Covariates

Prepare occupancy and spatial capture-recapture analyses, and summarise surveys

cameraOperation	Create a camera operability matrix

detectionHistory	Species detection histories for occupancy analyses (single and multi-season)

spatialDetectionHistory	Detection histories of individuals for spatial capture-recapture analyses

surveyReport	Create a report about camera trap surveys and species detections

surveyDashboard	Shiny dashboard for summarizing and analyzing camera trap survey data

Sample data

camtraps	Sample camera trap station information table

recordTableSample	Sample species record table

recordTableIndividualSample	Single-species record table with individual IDs

camtrapsMultiSeason	Sample multi season camera trap station information table

recordTableSampleMultiSeason	Sample multi season species record table

recordTableIndividualSampleMultiSeason	Single-species multi season record table with individual IDs

timeShiftTable	Sample camera trap time shift information

Vignettes

1. Organising raw camera trap images	2. Identifying species and individuals

3. Extracting Data from Camera Trapping Images and Videos	4. Data exploration and visualisation

5. Multi-species occupancy models

Author

Juergen Niedballa

Maintainer:Juergen Niedballa <camtrapr@gmail.com>

References

Niedballa, J., Sollmann, R., Courtiol, A., Wilting, A. (2016): camtrapR: an R package for efficient camera trap data management. Methods in Ecology and Evolution, 7(12). https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.12600

camtrapR Google Group https://groups.google.com/forum/#!forum/camtrapr

Phil Harvey's ExifTool https://exiftool.org/

See also