Create a community (multi-species) occupancy model for JAGS or Nimble

Flexibly creates complete code and input data for community occupancy models for in JAGS amd Nimble, and automatically sets initial values and parameters to monitor. Supports fixed and random effects of covariates on detection and occupancy probabilities, using both continuous and categorical covariates (both site and site-occasion covariates).

Optionally includes data augmentation (fully open community, or up to known maximum number of species, or no data augmentation). Allows combination of all these parameters for fast and flexible customization of community occupancy models.

Incidentally, the function can also be used to create model code and input for single-species single-season occupancy models (it is the special case with only one species). Model will run slower than proper single-species model JAGS code due to the additional species loop, but it is possible.

The function returns several derived quantities, e.g. species richness, Bayesian p-values (overall and by species), Freeman-Tukey residuals for actual and simulate data (by station and total). If doing data augmentation, metacommunity size and number of unseen species are returned also.

communityModel(
  data_list,
  occuCovs = list(fixed = NULL, ranef = NULL),
  detCovs = list(fixed = NULL, ranef = NULL),
  detCovsObservation = list(fixed = NULL, ranef = NULL),
  effortCov = "effort",
  intercepts = list(det = "fixed", occu = "fixed"),
  richnessCategories = NULL,
  augmentation = NULL,
  modelFile = NULL,
  nimble = FALSE
)

Arguments

data_list
list. Contains 3 slots: ylist, siteCovs, obsCovs. ylist ist a list of detection histories (can be named), e.g. from detectionHistory. siteCovs is a data.frame with site covariates (optional). obsCovs is a list of site-occasion level covariates (e.g. site-occasion-specific effort, which is also returned by detectionHistory.

occuCovs
list. Up to 2 items named "fixed" and/or "ranef". Specifies fixed or random effects of covariates on occupancy probability (continuous or categorical covariates)

detCovs
list. Up to 2 items named "fixed" and/or "ranef". Specifies fixed or random effects of covariates on detection probability (continuous or categorical covariates)

detCovsObservation
list. Up to 2 items named "fixed" and/or "ranef". Specifies fixed or random effects of observation-level covariates on detection probability (continuous or categorical covariates - categorical must be coded as character matrix)

effortCov
character. Name of list item in data_list$obsCovs which contains effort. This does not include effort as a covariate on detection probability, but only uses NA / not NA information to create binary effort and ensure detection probabilities p are 0 when there was no effort (p will be 0 whereever effortCov is NA).

intercepts
list. For detection and occupancy probability intercepts, are they "fixed" (= constant across species) or "ranef" (= random effect of species on intercept)?

richnessCategories
character. Name of categorical covariate in data_list$siteCovs for which to calculate separate richness estimates (optional). Can be useful to obtain separate richness estimates for different areas.

augmentation
If NULL, no data augmentation (only use species in data_list$ylist), otherwise named list or vector with total number of (potential) species. Names: "knownmax" or "full". Example: augmentation = c(knownmax = 30) or augmentation = c(full = 30)

modelFile
character. Text file name to save model to

nimble
logical. If TRUE, model code will be for Nimble (incompatible with JAGS). If FALSE, model code is for JAGS.

Value

commOccu object. It is an S4 class with the following slots:

modelText

JAGS model code as a character vector (made up of code chunks, use cat() to print)

params

Parameters to monitor in the model runs

inits_fun

Function to create start values for the MCMC chains. It being a function ensures different values in each chain

data

List with data needed to run the model (detection & effort matrices, site covariates, number of species / stations / occasions)

input

Input data_list (unchanged)

nimble

logical indicator for whether it is a Nimble model

modelFile

Path of the text file containing the model code

See commOccu-class
Details

For examples of implementation, see Vignette XXX.

Fixed effects of covariates are constant across species, whereas random effect covariates differ between species. Fixed and random effects are allowd for station-level detection and occupancy covariates (a.k.a. site covariates), and also station-occasion level covariates (a.k.a. observation covariates).

By default, random effects will be by species. It is however possible to use categorical site covariates for grouping (continuous|categorical). Furthermore, is is possible to use use nested random effects of species and another categorical site covariate (so that there is a random effect of species and an additional random effect of a categorical covariate within each species).

Derived quantities returned by the model are:

Bpvalue	Bayesian p-value (overall)

Bpvalue_species	Bayesian p-value (by species)

Nspecies	Species richness (only in JAGS model)

Nspecies_Covariate	Species richness by categorical covariate (when using richnessCategories, only in JAGS model)

R2	sum of Freeman-Tukey residuals of observed data within each species

new.R2	sum of Freeman-Tukey residuals of simulated data within each species

R3	Total sum of Freeman-Tukey residuals of observed data

new.R3	Total sum of Freeman-Tukey residuals of simulated data

Ntotal	Total metacommunity size (= observed species + n0)

n0	Number of unseen species in metacommunity

omega	Data augmentation parameter

w	Metacommunity membership indicator for each species

Quantities in italic at the bottom are only returned in full data augmentation. Nspecies and Nspecies_Covariate are only returned in JAGS models (because Nimble models don't explicitly return latent occupancy status z).

Parameter naming convention

The parameter names are assembled from building blocks. The nomenclature is as follows:

Type	Name	Description

Submodel	alpha	detection submodel

Submodel	beta	occupancy submode

Intercept	0	denotes the intercepts (alpha0, beta0)

Effect type	fixed	fixed effects (constant across species)

Effect type	ranef	random effects (of species and/or other categorical covariates)

Covariate type	cont	continuous covariates

Covariate type	categ	categorical covariates

Hyperparameter	mean	mean of random effect

Hyperparameter	sigma	standard deviation of random effect

Hyperparameter	tau	precision of random effect (used internally, not returned)

For example, a fixed intercept of occupancy (constant across species) is beta0.

An intercept with a random effect of species is:

beta0.mean community mean of the occupancy probability intercept (= logit(occupancy probability) at covariate values of 0).

beta0.sigma standard deviation of the estimated occupancy probability intercept.

beta0[1] occupancy intercept of species 1 (cont. for other species).

For effects of site covariates, the patters is:

submodel.effectType.covariateType.CovariateName.hyperparameter

For example:

beta.ranef.cont.habitat.mean is the mean community effect of the continuous site covariate 'habitat' on occupancy probability

beta.ranef.cont.habitat[1] is the effect of continuous site covariate 'habitat' on occupancy probability of species 1

Site-occasion covariates are denoted by ".obs" after the submodel, e.g.:

alpha.obs.fixed.cont.effort is the fixed effect (constant across species) of the continuous observation-level covariate 'effort' on detection probability

References

Kéry, M., and J. A. Royle. "Applied hierarchical modelling in ecology - Modeling distribution, abundance and species richness using R and BUGS." Volume 1: Prelude and Static Models. Elsevier/Academic Press, 2016.

Author

Juergen Niedballa

Examples

if (FALSE) {

data("camtraps")

# create camera operation matrix
camop_no_problem <- cameraOperation(CTtable      = camtraps,
                                    stationCol   = "Station",
                                    setupCol     = "Setup_date",
                                    retrievalCol = "Retrieval_date",
                                    hasProblems  = FALSE,
                                    dateFormat   = "dmy"
)

data("recordTableSample")

# list of detection histories
DetHist_list <- lapply(unique(recordTableSample$Species), FUN = function(x) {
  detectionHistory(
    recordTable         = recordTableSample,
    camOp                = camop_no_problem,
    stationCol           = "Station",
    speciesCol           = "Species",
    recordDateTimeCol    = "DateTimeOriginal",
    species              = x,
    occasionLength       = 7,
    day1                 = "station",
    datesAsOccasionNames = FALSE,
    includeEffort        = TRUE,
    scaleEffort          = TRUE,
    timeZone             = "Asia/Kuala_Lumpur"
  )}
)

names(DetHist_list) <- unique(recordTableSample$Species)

ylist <- lapply(DetHist_list, FUN = function(x) x$detection_history)

# some fake covariates for demonstration
sitecovs <- camtraps[, c(1:3)]
sitecovs$elevation <- c(300, 500, 600)   
sitecovs[, c(2:4)] <- scale(sitecovs[,-1])   # scale numeric covariates

# bundle input data for communityModel
data_list <- list(ylist = ylist,
                  siteCovs = sitecovs,
                  obsCovs = list(effort = DetHist_list[[1]]$effort))

# create community model for JAGS
modelfile1 <- tempfile(fileext = ".txt")
mod.jags <- communityModel(data_list,
                           occuCovs = list(fixed = "utm_y", ranef = "elevation"),
                           detCovsObservation = list(fixed = "effort"),
                           intercepts = list(det = "ranef", occu = "ranef"),
                           modelFile = modelfile1)

summary(mod.jags)

# fit in JAGS
fit.jags <- fit(mod.jags,
                n.iter = 1000,
                n.burnin = 500,
                chains = 3)   
summary(fit.jags)

# response curves (= marginal effect plots)
plot_effects(mod.jags, 
             fit.jags, 
             submodel = "state")
plot_effects(mod.jags, 
             fit.jags, 
             submodel = "det")
             
# effect sizes plot
plot_coef(mod.jags, 
          fit.jags, 
          submodel = "state")
plot_coef(mod.jags, 
          fit.jags, 
          submodel = "det")              

# create community model for Nimble
modelfile2 <- tempfile(fileext = ".txt")
mod.nimble <- communityModel(data_list,
                             occuCovs = list(fixed = "utm_x", ranef = "utm_y"),
                             detCovsObservation = list(fixed = "effort"),
                             intercepts = list(det = "ranef", occu = "ranef"),
                             modelFile = modelfile2, 
                             nimble = TRUE)      # set nimble = TRUE

# load nimbleEcology package 
# currently necessary to do explicitly, to avoid additional package dependencies
require(nimbleEcology)

# fit uncompiled model in Nimble
fit.nimble.uncomp <- fit(mod.nimble, 
                         n.iter = 10, 
                         chains = 1)

# fit compiled model in Nimble
fit.nimble.comp <- fit(mod.nimble, 
                       n.iter = 5000, 
                       n.burnin = 2500,
                       chains = 3, 
                       compile = TRUE)

# parameter summary statistics
summary(fit.nimble.comp)

# response curves (= marginal effect plots)
plot_effects(mod.nimble, 
             fit.nimble.comp, 
             submodel = "state")
plot_effects(mod.nimble, 
             fit.nimble.comp, 
             submodel = "det")

#' # effect sizes plot
plot_coef(mod.nimble, 
          fit.nimble.comp, 
          submodel = "state")
plot_coef(mod.nimble, 
          fit.nimble.comp, 
          submodel = "det")   

# traceplots
plot(fit.nimble.comp)

}