Shiny dashboard for summarizing and analyzing camera trap survey data

Open a Shiny dashboard to summarize a camera trapping survey. It provides an overview of the data, maps, can create and run occupancy models in unmarked, and plot species activity.

Maps are interactive and can be panned, zoomed and queried. Users can choose a suitable basemap in the layers button.

For occupancy models, users can flexibly customize detection histories. Then the user can specify the model structure (by selecting site covariates on detection and occupancy probability from the camera trap table) and optionally including effort. Models are fitted automatically with unmarked. Response curves (marginal effect plots), model and parameter summaries are computed and updated automatically with every change to the model. Multiple models can be compared using model selection.

surveyDashboard(
  CTtable,
  recordTable,
  stationCol,
  cameraCol = NULL,
  xcol,
  ycol,
  crs,
  setupCol,
  retrievalCol,
  hasProblems = FALSE,
  CTdateFormat = "ymd",
  camerasIndependent,
  speciesCol = "Species",
  recordDateTimeCol = "DateTimeOriginal",
  recordDateTimeFormat = "ymd HMS",
  timeZone = "UTC",
  exclude = NULL
)

Arguments

CTtable

A data.frame containing the camera trap deployment information.

recordTable

A data.frame containing the camera trap records.

stationCol

The column name containing the camera trap station ID

cameraCol

The column name containing the camera trap IDs (optional, only if 2 or more cameras per station)

xcol

The column name containing the X coordinate of the camera trap station.

ycol

The column name containing the Y coordinate of the camera trap station.

crs

The coordinate reference system (CRS) of the camera trap data. Must be a valid argument to st_crs

setupCol

The column name containing the camera trap deployment date (and time).

retrievalCol

The column name containing the camera trap retrieval date (optionally date-time).

hasProblems

A logical indicating whether there are periods of cameras malfunctioning

CTdateFormat

The date format of the camera trap deployment and retrieval date and time (default: "ymd").

camerasIndependent

logical. If multiple camera per station, are they independent?

speciesCol

The column name containing the species names

recordDateTimeCol

The column name containing the record date and time

recordDateTimeFormat

The date/time format of recordDateTimeCol

timeZone

Time zone of records in recordTable

exclude

Species to be excluded from the data set

Value

A Shiny dashboard for camera trap survey data.

Note

Current limitations include:

- supports only single-season data

- only single-species occupancy models

- no random or interaction effects in occupancy models

- no (spatial) model predictions yet (planned for Q4/2023)

- no support for spatial capture-recapture models (or anything related to individual IDs)

- only models in unmarked, no Bayesian models in ubms yet (likely to be included later)

Author

Juergen Niedballa

Examples

if (FALSE) {
data("camtraps")
data("recordTableSample")
 
  surveyDashboard(
    CTtable = camtraps,
    recordTable = recordTableSample,
    xcol = "utm_x",
    ycol = "utm_y",
    crs = "epsg:32650",      # = UTM50N (Sabah, Malaysian Borneo)
    stationCol = "Station",
    setupCol = "Setup_date",
    retrievalCol = "Retrieval_date",
    CTdateFormat = "dmy"
  )
  }